# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EventHub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List published events (public)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Events list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event (organizer only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateEvent
      tags: [events]
      summary: Update event (organizer only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete event (organizer only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /events/{id}/attendance:
    get:
      operationId: getEventAttendance
      tags: [events]
      summary: Live attendance dashboard for an event (organizer)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Attendance data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceDashboard"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/{id}/analytics:
    get:
      operationId: getEventAnalytics
      tags: [events]
      summary: Analytics for an event (organizer)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventAnalytics"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Event:
      type: object
      required: [id, title, category, venue, startTime, endTime, capacity, status, organizerId, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        bannerUrl:
          type: ["string", "null"]
        venue:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        capacity:
          type: integer
        registrationDeadline:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          enum: [draft, published, cancelled, archived]
        organizerId:
          type: integer
        registeredCount:
          type: integer
        checkedInCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    EventDetail:
      type: object
      required: [id, title, category, venue, startTime, endTime, capacity, status, organizerId, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        bannerUrl:
          type: ["string", "null"]
        venue:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        capacity:
          type: integer
        registrationDeadline:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          enum: [draft, published, cancelled, archived]
        organizerId:
          type: integer
        organizerName:
          type: string
        registeredCount:
          type: integer
        checkedInCount:
          type: integer
        isRegistered:
          type: boolean
        createdAt:
          type: string
          format: date-time
    EventListResponse:
      type: object
      required: [events, total, page, limit]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    EventInput:
      type: object
      required: [title, category, venue, startTime, endTime, capacity]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        category:
          type: string
        bannerUrl:
          type: string
        venue:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        capacity:
          type: integer
          minimum: 1
        registrationDeadline:
          type: string
          format: date-time
        status:
          type: string
          enum: [draft, published]
    EventUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        category:
          type: string
        bannerUrl:
          type: string
        venue:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        capacity:
          type: integer
          minimum: 1
        registrationDeadline:
          type: string
          format: date-time
        status:
          type: string
          enum: [draft, published, cancelled, archived]

    # ─── Registrations ────────────────────────────────────────────────────────
...
    AttendanceDashboard:
      type: object
      required: [eventId, totalRegistered, totalCheckedIn, totalNoShow, recentCheckins]
      properties:
        eventId:
          type: integer
        totalRegistered:
          type: integer
        totalCheckedIn:
          type: integer
        totalNoShow:
          type: integer
        recentCheckins:
          type: array
          items:
            $ref: "#/components/schemas/CheckinLogEntry"
    CheckinLogEntry:
      type: object
      required: [id, attendeeName, action, timestamp]
      properties:
        id:
          type: integer
        attendeeName:
          type: string
        attendeeEmail:
          type: string
        action:
          type: string
          enum: [check_in, check_out]
        station:
          type: ["string", "null"]
        scannedByName:
          type: ["string", "null"]
        timestamp:
          type: string
          format: date-time

    # ─── Analytics ────────────────────────────────────────────────────────────
    EventAnalytics:
      type: object
      required: [eventId, totalRegistered, totalCheckedIn, capacityUtilization, registrationTrend, checkinTimeline]
      properties:
        eventId:
          type: integer
        totalRegistered:
          type: integer
        totalCheckedIn:
          type: integer
        totalWaitlisted:
          type: integer
        totalCancelled:
          type: integer
        capacityUtilization:
          type: number
        checkinRate:
          type: number
        registrationTrend:
          type: array
          items:
            $ref: "#/components/schemas/TimeSeriesPoint"
        checkinTimeline:
          type: array
          items:
            $ref: "#/components/schemas/TimeSeriesPoint"
        volunteerTaskCompletion:
          type: number
    TimeSeriesPoint:
      type: object
      required: [date, count]
      properties:
        date:
          type: string
        count:
          type: integer

    # ─── Volunteers ───────────────────────────────────────────────────────────
...